requirements.txt
PyStemmer>=1.3,<2.0
numpy>=1.12,<2.0
scipy>=0.17,<1.0
tensorflow>=1.0,<2.0
clint>=0.5.0
asdf>=1.2,<2.0
google-cloud-storage>=1.0,<2.0
python-dateutil
lz4>=0.10.1
modelforge>=0.2.3-alpha
bblfsh>=0.0.4
